requirements.txt
# Essential LangChain core
langchain-core==0.3.74

# Modern LangGraph for agents (includes langchain dependencies)
langgraph==0.2.68

# OpenAI integration
langchain-openai==0.2.11

# Environment variable management
python-dotenv==1.1.1

# Database for TimeMachine storage
sqlalchemy==2.0.43

# Web API backend for Phase 3
fastapi==0.104.1
uvicorn==0.24.0